+++ packages/core/plot/src/Roster.js
@@ -1,6 +1,6 @@
-import { presFlopper, presShifter } from '@palett/flopper'
-import { decoString }               from '@spare/deco-string'
-import { hasAnsi }                  from '@texting/charset-ansi'
+import { presFlopper, presShifter, rhodFlopper } from '@palett/flopper'
+import { decoString }                            from '@spare/deco-string'
+import { hasAnsi }                               from '@texting/charset-ansi'
 
 
 export class Roster {
@@ -13,8 +13,11 @@ export class Roster {
 
   static build(presGen) { return new Roster(presGen) }
   static from(book, mode) {
-    const makeGen = mode === 'flop' ? presFlopper : mode === 'shift' ? presShifter : presFlopper
-    return new Roster(makeGen.call({ pres: book }))
+    const makeGen =
+      mode === 'rhod' ? rhodFlopper
+        : mode === 'shift' ? presShifter
+          : presFlopper
+    return new Roster(makeGen.call(book))
   }