+++ packages/core/plot/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @spare/xr
 
+## 0.16.19
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-string@0.16.19
+
+## 0.16.18
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-string@0.16.18
+
 ## 0.16.17
 
 ### Patch Changes